feat(authmiddleware): add authorization middleware
This middleware checks for the Authorization header for the JWT token. If empty, returns not allowed
error. If not empty but invalid, tries to refresh the tokens and sends it to the request context.
Then, the controller should let the user know that the tokens are changed, changing the return
headers